requirements.txt
# Core conversion dependencies
openpyxl==3.1.5
numpy==2.3.4
pandas==2.3.3

# Web application dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
gitpython==3.1.41
